Reverted r112214, since it was not the right fix for the build.
* css/StylePropertySet.cpp:
(WebCore::StylePropertySet::asText):



Canonical link: https://commits.webkit.org/99657@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@112220 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/css/StylePropertySet.cpp
@@ -716,12 +716,12 @@ String StylePropertySet::asText() const
                 continue;
             if (!shorthandPropertyAppeared.get(shortPropertyIndex))
                 value = getPropertyValue(shorthandPropertyID);
-            shorthandPropertyAppeared.ensureSizeAndSet(static_cast<size_t>(shortPropertyIndex), static_cast<size_t>(numCSSProperties));
+            shorthandPropertyAppeared.ensureSizeAndSet(shortPropertyIndex, numCSSProperties);
         }
 
         if (!value.isNull()) {
             propertyID = shorthandPropertyID;
-            shorthandPropertyUsed.ensureSizeAndSet(static_cast<size_t>(shortPropertyIndex), static_cast<size_t>(numCSSProperties));
+            shorthandPropertyUsed.ensureSizeAndSet(shortPropertyIndex, numCSSProperties);
         } else
             value = prop.value()->cssText();
